Release 1.3.1 (#226)
* #I345: fix:automted private and public key generation

* #I345: fix:automated the keycloak helm chart

* #I345: fix:automated the keycloak helm chart

* #I345: fix:updated the keys file name

* #I345: fix:updated the keycloak database creation script

* #I345: fix:updated the .gitignore

* #I345: fix:updated the dataset-api and webconole config

* #I345: fix:updated the dataset-api and webconole config

* #OBS-I345 keycloak changes

* #OBS-I345 update letsencrypt bundle

* #OBS-I345 keycloak ingress and other changes

* OBS-I363 : fix:velero backup using service account

* Kafka exporter vulnerability fixes (#218)

* #I431: updated node js version

* #I431:  Removed unused package

* #I345: fix: private and public keys autoamtion and keycloak chanages

* #I345: fix: updated resource for keycloak service

* I345: fix:updated the configuration

* Critical Vulnerabilities Fixes (#216)

* #I431: 1. Upgraded the superset from 3.0.2 to 4.1.1
2. Velero Upgrade from 1.10.2 to 1.15.0
3. postgresql-exporter upgrade from v0.11.1 to v0.16.0
4. Grafana upgrade from 9.5.2 to 11.4.0

* #I431: 1.Upgrade the promtail from 2.7.4 to 3.2.2
2. Upgraded alertmanage from v0.25.0 to v0.28.0-rc.0
3. Upgraded the prometheus/node-exporter from v1.5.0 to v1.8.2
4. Upgraded the Postgres exporter from v0.11.1 to v0.16.0

* #I431: 1. Upgraded the grafana loki from 2.9.3 to 3.2.2

* #I431: 1. Upgraded the prom operator from v0.65.1 to v0.78.2

* #I431: 1. Upgraded the kube state metrics from 2.8.2 to 2.14.0

* #I431: 1. Upgraded the redis from 7.0.11-debian-11-r15 to 7.4.1

* #I431: 1. Upgraded the prometheus config reloader from v0.65.1 to v0.79.0 and nginx-unprivileged from 1.24-alpine to 1.27-alpine

* #I431: 1. Upgraded the prometheus config reloader from v0.65.1 to v0.79.0 and nginx-unprivileged from 1.24-alpine to 1.27-alpine

* #I431: 1. Resolved the vuln issues by upgrading certain dependencies

* #I431: 1. Vuln fixes

* #I431: 1.Upgrade the bats from v1.4.1 to 1.11.1

* #I431: Upgraded the golang in the druid exporter and disabled the grafana-testframework which is no longer needed

---------


* I345: fix:removed the dev.tfvars file

* I345: fix:removed debug option in install.sh

* OBS-I392: fix:updated private and public keys configuration

* OBS-I392: fix:updated private and public keys configuration

* OBS-I392: fix:updated private and public keys configuration

* #OBS-I345 disable keycloak auth until superset and grafana are configured

* #OBS-I345 feat: Add Dockerfile for Superset (#220)

* #OBS-I438 superset and other fixes

* #OBS-I438 superset and other fixes

* #OBS-I199 disable oauth local dc

* #OBS-I199 disable oauth in all

* #OBS-I345 add hosts only if ssl is enabled

* #OBS-I345 disable kong in local dc

* #OBS-I433 flink sa default to false

* #OBS-I433 flink with credentials

* #OBS-I433 flink with credentials

* #OBS-I433 flink remove incorrect vals

* #OBS-I345 update default keys

* #OBS-I437 upgrade loki helm chart

* #OBS-I437 upgrade loki values

* #OBS-I437 update loki configurations

* #OBS-I393: remove dev prefix from all the kafka topics wherevere configured.

* #OBS-I436 update install.sh to take args

* #OBS-I436 update values.yaml web-console

* #OBS-I436 update secret creation jobs

* #OBS-I448 automate creation of keys

* #OBS-I448 fix reflection ns

* #OBS-I436 update masterdata to cacheidx

* #OBS-I436 update postgres backup image

* #OBS-I345 feat: Add Dockerfile for Superset (#220)

* #OBS-I436 images

---------

+++ helmcharts/kitchen/install.sh
@@ -22,9 +22,9 @@ esac
 cp -rf ../{global-values.yaml,global-resource-values.yaml,images.yaml} ./
 
 if [ "$2" == "template" ]; then
-    cmd="template"
+    cmd="template ${@: 3}"
 else
-    cmd="upgrade -i "
+    cmd="upgrade -i ${@: 2}"
 fi
 
 case "$1" in
@@ -52,7 +52,7 @@ migrations)
 
     if [ -z "$cloud_env" ]; then
         cp -rf ../services/minio migrations/charts/
-    fi  
+    fi
 
     helm $cmd migrations ./migrations -n obsrv -f global-resource-values.yaml -f global-values.yaml -f images.yaml -f $cloud_file_name
     ;;
@@ -88,12 +88,15 @@ otel)
     helm $cmd opentelemetry-collector ./opentelemetry-collector -n obsrv -f global-resource-values.yaml -f global-values.yaml -f images.yaml -f $cloud_file_name
     ;;
 oauth)
-    rm -rf oauth
-    cp -rf ../obsrv oauth
-    cp -rf ../services/keycloak oauth/charts/
-    helm $cmd oauth ./oauth -n obsrv -f global-resource-values.yaml -f global-values.yaml -f images.yaml -f $cloud_file_name
+    if [ -z "$cloud_env" ]; then
+        echo "oauth not yet supported for local datacenter"
+    else
+        rm -rf oauth
+        cp -rf ../obsrv oauth
+        cp -rf ../services/keycloak oauth/charts/
+        helm $cmd oauth ./oauth -n obsrv -f global-resource-values.yaml -f global-values.yaml -f images.yaml -f $cloud_file_name
+    fi
     ;;
-
 obsrvtools)
     rm -rf obsrvtools
     cp -rf ../obsrv obsrvtools
@@ -118,22 +121,24 @@ additional)
     helm $cmd additional ./additional -n obsrv -f global-resource-values.yaml -f global-values.yaml -f images.yaml -f $cloud_file_name
     ;;
 all)
-    bash $0 bootstrap
-    bash $0 coredb
-    bash $0 migrations
-    bash $0 monitoring
-    bash $0 coreinfra
-    bash $0 obsrvapis
-    bash $0 hudi
-    bash $0 otel
-    bash $0 oauth
-    bash $0 obsrvtools
-    bash $0 additional
+    bash $0 bootstrap ${@: 2}
+    bash $0 coredb ${@: 2}
+    bash $0 migrations ${@: 2}
+    bash $0 monitoring ${@: 2}
+    bash $0 coreinfra ${@: 2}
+    bash $0 obsrvapis ${@: 2}
+    bash $0 hudi ${@: 2}
+    # bash $0 otel ${@: 2}
+    bash $0 oauth ${@: 2}
+    bash $0 obsrvtools ${@: 2}
+    bash $0 additional ${@: 2}
 
     ;;
 reset)
     helm uninstall additional -n obsrv
     helm uninstall obsrvtools -n obsrv
+    helm uninstall oauth -n obsrv
+    helm uninstall otel -n obsrv
     helm uninstall hudi -n obsrv
     helm uninstall obsrvapis -n obsrv
     helm uninstall coreinfra -n obsrv
